modules/compression: fix function calls
Some functions calls use an int* instead of a long int*, or an int
instead ok int*. Fixed by looking at the surrounding code, but knowing
the correct syntax is difficult. For the moment, the fix only focuses on
clean compilation.

+++ modules/src/compression/lossless/ardecode2.c
@@ -86,7 +86,7 @@ REALLOCATE_OUTPUT(Fimage output)
 
 
 static void
-START_MODEL(Fsignal histo, int *cap_histo, int predict)
+START_MODEL(Fsignal histo, long int * cap_histo, long predict)
 {
   int     y, z;
